Split config panels into separate components.
